NEW Consolidates REST dictionary APIs into a single tree and a single file
The php code of the files "htdocs / api / class / api_dictionary * .class.php" has been integrated into a single file: "htdocs / api / class / api_dictionary.class.php".
